Split up operations.rkt
The operations file would get too large over time.

Also, an update operations doesn't really make sense with a side-by-side
installation model, except as an abbreviation.

That's meant to be more of a "throw stuff in" experience, so updates are really
just a matter of installing a package and then trying to uninstall a prior one.