Refactor messaging model
I've been trying to manually coordinate
access to stdin between a parent process
and a bunch of places. That got old fast.

I realized that I could probably scale
up better if I used a flat list of defined
messages as events. This commit refactors
the code to that effect.