Refactor to use per-file configuration
I'm liking the crazy idea of one Racket value per config file,
and this commit refactors the project to do that instead of
using an info.rkt file. `cat`, etc. allows you to review an entire
configuration as a document anyway. Individual files are much
easier to read, update, and lock.