Draft refactor for configuration API
This marks the beginning of functional tests for
changes to package metadata, runtime configuration,
and both the `config` and `new` commands.

The problems I've been solving are as follows:

1. I experimented with a configuration approach with one file per value.  That
   was as crazy as expected I shouldn't have entertained it as much as I did.

2. Package metadata and runtime configuration have evolved independently.

3. Configuration has been loading as a side effect of module instantiation.

Here, the config.rkt and setting.rkt modules define a better basis
for rcfiles and loading values from a variety of sources. Both
the package manager and packages now load configuration from
files of the same format, and none of the package manager's
configuration loads until you actually enter the `main`
submodule of the CLI.

The settings are no longer managed by parameters, since the
macro magic around them got a little crazy. I needed a custom
structure type for settings, and once I had it, things got
a little cleaner.

Don't expect everything to work as of this commit. However,
the config.rkt and settings.rkt files are well tested.