Commit stopping point for team/worker refactor
Parallel programming is hard enough, but I'm starting to think that committing
to functional updates on structs was a bad move.  Message passing is
imperative, so mixing functional code made things really hard to follow. I
spent hours on cryptic errors that boiled down to using the wrong binding in
one spot.

Tests pass, except the CLI, which I have not updated.  I did not update the
tests because I don't think my refactoring is over. I might end up turning to
OOP for this problem domain, because at the end of the day I am trying to
encapsulate data and behavior.  Why fight that?

I just have to remember that the objects need to be easy to test. In this
commit I did manage to make a message passing model where messages are handled
in top level functions. That's easy to unit test, but those functions need to
return updated values for monolithic state objects. The trees were easy to
test, but the forest out to fucking get me.