Draft refactor of parallel build model
This yak is about fully shaved, but all modules compile as
of this commit. The actor.rkt package has 100% test coverage,
but zcpkg-actor.rkt does not. The latter uses project semantics
on top of actor.rkt's provided class model.

I have no idea if this will be the model to go on forever,
but the CLI certainly benefitted from the effort. It clearly
shows that whatever the package manager does is an interaction
between the user and an actor% that discovers and defers work.