Route all program output through output.rkt
This commit makes output.rkt responsible for almost all use of STDOUT.
This makes it easier to leverage settings for verbose output, coloring,
and different message formats.

Part of making this change work was refactoring cli.rkt. Each command
is now expected to reach a unique "halt" state that includes an
exit code and program output consisting of $message instances.
This increases the size of zcpkg-messages.rkt, but it means that
every event in the program has an explicitly declared event.

Another benefit to this approach is that functional tests do
not have to analyze human-readable text. zcpkg can be configured
to emit (read)able values as program output, and functional tests
can make assertions against those values instead.