Refactor approach to program output
Before this commit, all program output was accumulated
and then printed. This is not desireable because you
don't see output indicating a program's progress.

Still, accumulation is handy because it allows
a functional test to analyze output without
reading from buffers populated by STDOUT or STDERR.

This commit performs the following adjustments:

- output.rkt now stores a module-level buffer for accumulated output.
  You can get the output while optionally resetting the buffer, similar
  to get-output-bytes or get-output-string.

- cli.rkt no longer expects both an exit code and a list of program output.
  A command line handler must now produce only an exit code. Program output
  is written as a side effect, but is still accumulated by entry-point