Draft refactor of capture/restore
The restore command did not act like a rollback. It would only
add files. Since rollback semantics appear to be in demand,
this commit includes the ability to delete extra or changed
files before recreating files in the workspace.

The restore command no longer runs a diff command at
the end. It instead runs a diff in advance and uses
the output to make decisions. This comes with new
messages and scope of work output for the user.

This commit includes a refactor to bring more code into capture.rkt.