Include openssl stdout in related exception
OpenSSL sometimes returns a 1 exit code for successfully computing
an answer, such as when verifying a signature and finding a mismatch.
This is a quick patch that includes STDOUT in the raised exception
so that a caller can handle errors that aren't actually exceptional.

This probably is not the best solution. In the future, consider letting
callers try a different convention for starting an OpenSSL subprocess.
Or, better yet, use a FFI.